java/src/main/java/com/yourtion/leetcode/daily/m09/d28/Solution.java
package com.yourtion.leetcode.daily.m09.d28;

import com.yourtion.leetcode.utils.binary.Node;

import java.util.LinkedList;
import java.util.Queue;

/**
 * 117. 填充每个节点的下一个右侧节点指针 II
 *
 * @author Yourtion
 * @link https://leetcode-cn.com/problems/populating-next-right-pointers-in-each-node-ii/
 */
public class Solution {
    public Node connect(Node root) {
        if (root == null) {
            return null;
        }
        Queue<Node> queue = new LinkedList<>();
        queue.offer(root);
        while (!queue.isEmpty()) {
            int n = queue.size();
            Node last = null;
            for (int i = 1; i <= n; ++i) {
                Node f = queue.poll();
                if (f == null) {
                    continue;
                }
                if (f.left != null) {
                    queue.offer(f.left);
                }
                if (f.right != null) {
                    queue.offer(f.right);
                }
                if (i != 1) {
                    last.next = f;
                }
                last = f;
            }
        }
        return root;
    }
}
